Add OnPageChangeListener - Fix nav flow error (#42)
* Add OnPageChangeListener - Fix nav flow error

This bug occurs when you re-navigate to a tabbed fragment  that now has some text in it, and thus should display the next arrow.

EX: Go forward a few times, then go backwards to Name Contest…and then forward again to Describe Contest Page

* Add OnPageChangeListener - Fix nav flow error

* Move OnPageChangeListener to Presenter

based on review. Did not have any luck with attempting to leverage the presenter.OnViewBound callback. I think the viewpager fragment lifecycles is not as straightforward as with hosting a single fragment.